fix type warning

+++ src/oplist.c
@@ -318,7 +318,7 @@ void dumpTo (struct OpList *self, FILE *file)
 			fprintf(file, "}\n");
 		}
 		else if (self->ops[i].native == Op.getParentSlot || self->ops[i].native == Op.getParentSlotRef || self->ops[i].native == Op.setParentSlot)
-			fprintf(file, "[-%d] %d", self->ops[i].value.data.integer >> 16, self->ops[i].value.data.integer & 0xffff);
+			fprintf(file, "[-%hu] %hu", (uint16_t)(self->ops[i].value.data.integer >> 16), (uint16_t)(self->ops[i].value.data.integer & 0xffff));
 		else if (self->ops[i].value.type != Value(undefinedType) || self->ops[i].native == Op.value || self->ops[i].native == Op.exchange)
 			Value.dumpTo(self->ops[i].value, file);
